backend/requirements.txt
fastapi
uvicorn
openai
pandas
numpy
httpx
python-dotenv
langchain
langchain-cohere
langchain-community
langchain-core
langchain-huggingface
langchain-openai
langchain-text-splitters
pyarrow
PyMuPDF
pypdf
faiss-cpu
transformers
tiktoken
sentence_transformers
pypdf
